Refactor Event Instance & Improve Auto-initialization
FmodEventInstance now uses RefCounted instead of Node. This allows for the instance's internal fmod representation to be automatically released when no longer needed. Because of this change, FmodManager::oneshot was removed, since the event instance will automatically call release when the instance goes out of scope.

Removed bloat and unnecessary methods. Rewrote auto-initialization and singleton handling.